% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jirar.R
\docType{package}
\name{jira}
\alias{jira}
\alias{jira-package}
\title{Connectivity to 'JIRA' Issue Tracking Software}
\description{
Allows you to connect to 'JIRA' issue tracking software, gather data about your projects and analyis the data using 'R'.
\itemize{
\item \code{\link{jirar_search}} - Search for issues in JIRA
}
}
\author{
Roy Wetherall \email{rwetherall@gmail.com}
}
