% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jirar_query.R
\name{jirar_query}
\alias{jirar_query}
\title{Execute a JQL query}
\usage{
jirar_query(server, user, password, query, max_results = 1000)
}
\arguments{
\item{server}{server URL for JIRA}

\item{user}{user name}

\item{password}{password}

\item{query}{JQL query string}

\item{max_results}{maximum number of items returned in the result set}
}
\value{
Results of the query
}
\description{
Executes a JQL query and processes the results.
}
