% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jirar_search.R
\name{jirar_search}
\alias{jirar_search}
\title{Search for issues in JIRA}
\usage{
jirar_search(server, user, password, jql, start_at = 0,
  max_results = 1000)
}
\arguments{
\item{server}{server URL for JIRA}

\item{user}{user name}

\item{password}{password}

\item{jql}{JQL query string}

\item{start_at}{starting position for results}

\item{max_results}{maximum number of items returned in the result set}
}
\value{
Results of the query
}
\description{
Uses a JQL query to search for issues within JIRA.
}
